+++ app/src/common/docs/BaseDocsBlock.tsx
@@ -85,6 +85,7 @@ export abstract class BaseDocsBlock extends React.Component<any, BaseDocsBlockSt
             key: 'value',
             caption: 'Type',
             render: prop => <Text color='gray80'><span style={ { whiteSpace: "pre-wrap" } }>{ prop.value }</span></Text>,
+            width: 200,
             grow: 1,
             isSortable: true,
         },